requirements.txt
acme==0.31.0
APScheduler==3.6.3
asn1crypto==0.24.0
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certbot==0.27.0
certbot-nginx==0.23.0
certifi==2020.12.5
chardet==4.0.0
click==8.0.1
ConfigArgParse==0.11.0
configobj==5.0.6
cryptography==2.1.4
dataclasses==0.8
distro-info===0.18ubuntu0.18.04.1
dokuwiki==1.3.2
Flask==2.0.1
future==0.15.2
google-api-core==2.0.0
google-api-python-client==2.18.0
google-auth==2.0.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.6
importlib-metadata==4.6.4
iotop==0.6
itsdangerous==2.0.1
Jinja2==3.0.1
josepy==1.1.0
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==2.0.1
mock==2.0.0
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
pandas==1.1.5
parsedatetime==2.4
pbr==3.1.1
protobuf==3.17.3
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyGObject==3.26.1
pyOpenSSL==17.5.0
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.2
python-telegram-bot==13.7
python-wordpress-xmlrpc==2.3
pytz==2021.1
pyxdg==0.25
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.8.0
rsa==4.7.2
SecretStorage==2.3.1
six==1.11.0
soupsieve==2.3.2.post1
tornado==6.1
typing-extensions==3.10.0.0
tzlocal==2.1
ufw==0.36
uritemplate==3.0.1
urllib3==1.26.4
websockets==9.1
Werkzeug==2.0.1
zipp==3.5.0
zope.component==4.3.0
zope.event==4.2.0
zope.hookable==4.0.4
zope.interface==4.3.2
